Truly move over the manager to the single module
