--- manager/context/applicationContext.go
@@ -1,16 +1,0 @@
-package context
-
-import (
-	"github.com/golobby/container"
-)
-
-
-type ApplicationContext struct{}
-
-func(appContext ApplicationContext) Bind(bindFunction interface{}) {
-	container.Singleton(bindFunction)
-}
-
-func(appContext ApplicationContext) Resolve(toResolve interface{}) {
-	container.Make(toResolve)
-}
