--- manager/context/gatewayInterfaces.go
@@ -1,5 +1,0 @@
-package context
-
-type EnvironmentGateway interface {
-	GetEnvironmentValue(key string) string
-}
