--- manager/context/libraryInterfaces.go
@@ -1,5 +1,0 @@
-package context
-
-type UuidLibrary interface {
-	GenerateRandomUuid() string
-}
