--- manager/external/lib/errors/pkgErrors.go
@@ -1,19 +1,0 @@
-package errors
-
-import "github.com/pkg/errors"
-
-func New(message string) error {
-	return errors.New(message)
-}
-
-func Wrap(err error, message string) error {
-	return errors.Wrap(err, message)
-}
-
-func Is(err, target error) bool {
-	return errors.Is(err, target)
-}
-
-func As(err error, target interface{}) bool {
-	return errors.As(err, target)
-}
