--- manager/external/lib/googleUuid.go
@@ -1,12 +1,0 @@
-package lib
-
-import (
-	"github.com/google/uuid"
-)
-
-type GoogleUuid struct {}
-
-func (receiver GoogleUuid) GenerateRandomUuid() string {
-	randomUuid := uuid.New()
-	return randomUuid.String()
-}
