--- manager/external/rest.go
@@ -1,65 +1,0 @@
-package external
-
-import (
-	"github.com/gin-gonic/gin"
-	"github.com/halprin/email-conceal/manager/context"
-	"github.com/halprin/email-conceal/manager/controllers"
-)
-
-var applicationContext = context.ApplicationContext{}
-
-func Rest() {
-	router := gin.Default()
-
-	v1 := router.Group("/v1")
-	v1.POST("/concealEmail", createConcealEmail)
-	v1.PUT("/concealEmail/:concealEmailId", updateConcealEmail)
-	v1.DELETE("/concealEmail/:concealEmailId", deleteConcealEmail)
-
-	_ = router.Run(":8000")
-}
-
-func createConcealEmail(context *gin.Context) {
-	var genericMap map[string]interface{}
-
-	err := context.BindJSON(&genericMap)
-	if err != nil {
-		return
-	}
-
-	var concealEmailController controllers.ConcealEmailController
-	applicationContext.Resolve(&concealEmailController)
-	httpStatus, jsonMap := concealEmailController.Add(genericMap)
-
-	context.JSON(httpStatus, jsonMap)
-}
-
-func deleteConcealEmail(context *gin.Context) {
-	concealEmailId := context.Param("concealEmailId")
-
-	requestMap := map[string]interface{} {
-		"concealEmailId": concealEmailId,
-	}
-
-	var concealEmailController controllers.ConcealEmailController
-	applicationContext.Resolve(&concealEmailController)
-	httpStatus, jsonMap := concealEmailController.Delete(requestMap)
-
-	context.JSON(httpStatus, jsonMap)
-}
-
-func updateConcealEmail(context *gin.Context) {
-	var requestMap map[string]interface{}
-	err := context.BindJSON(&requestMap)
-	if err != nil {
-		return
-	}
-
-	requestMap["concealEmailId"] = context.Param("concealEmailId")
-
-	var concealEmailController controllers.ConcealEmailController
-	applicationContext.Resolve(&concealEmailController)
-	httpStatus, jsonMap := concealEmailController.Update(requestMap)
-
-	context.JSON(httpStatus, jsonMap)
-}
