--- manager/gateways/awsSession.go
@@ -1,5 +1,0 @@
-package gateways
-
-import "github.com/aws/aws-sdk-go/aws/session"
-
-var awsSession, sessionErr = session.NewSession()
