--- manager/gateways/osEnvEnvironmentGateway.go
@@ -1,12 +1,0 @@
-package gateways
-
-import (
-	"os"
-)
-
-
-type OsEnvironmentGateway struct {}
-
-func (receiver OsEnvironmentGateway) GetEnvironmentValue(key string) string {
-	return os.Getenv(key)
-}
