--- manager/go.mod
@@ -1,11 +1,0 @@
-module github.com/halprin/email-conceal/manager
-
-go 1.14
-
-require (
-	github.com/aws/aws-sdk-go v1.28.13
-	github.com/gin-gonic/gin v1.6.3
-	github.com/golobby/container v1.3.0
-	github.com/google/uuid v1.1.1
-	github.com/pkg/errors v0.9.1
-)
