--- manager/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	"github.com/halprin/email-conceal/manager/external"
-)
-
-func main() {
-	external.Rest()
-}
