--- manager/usecases/gatewayInterface.go
@@ -1,7 +1,0 @@
-package usecases
-
-type ConcealEmailGateway interface {
-	AddConcealedEmailToActualEmailMapping(concealPrefix string, actualEmail string, description *string) error
-	DeleteConcealedEmailToActualEmailMapping(concealPrefix string) error
-	UpdateConcealedEmail(concealPrefix string, description *string) error
-}
