+++ src/controllers/concealEmail.go
@@ -2,10 +2,10 @@ package controllers
 
 import (
 	"fmt"
-	"github.com/halprin/email-conceal/manager/context"
-	"github.com/halprin/email-conceal/manager/entities"
-	"github.com/halprin/email-conceal/manager/external/lib/errors"
-	"github.com/halprin/email-conceal/manager/usecases"
+	"github.com/halprin/email-conceal/src/context"
+	"github.com/halprin/email-conceal/src/entities"
+	"github.com/halprin/email-conceal/src/external/lib/errors"
+	"github.com/halprin/email-conceal/src/usecases"
 	"log"
 	"net/http"
 )
