+++ src/go.mod
@@ -7,7 +7,6 @@ require (
 	github.com/gin-gonic/gin v1.6.3
 	github.com/golobby/container v1.3.0
 	github.com/google/uuid v1.1.1
-	github.com/halprin/email-conceal/manager v0.0.0-20201108192054-433df20a28d4
 	github.com/pkg/errors v0.9.1
 	golang.org/x/net v0.0.0-20201216054612-986b41b23924 // indirect
 )
