+++ src/usecases/concealEmail.go
@@ -2,9 +2,9 @@ package usecases
 
 import (
 	"fmt"
-	"github.com/halprin/email-conceal/manager/context"
-	"github.com/halprin/email-conceal/manager/entities"
-	"github.com/halprin/email-conceal/manager/external/lib/errors"
+	"github.com/halprin/email-conceal/src/context"
+	"github.com/halprin/email-conceal/src/entities"
+	"github.com/halprin/email-conceal/src/external/lib/errors"
 )
